chore(tsdown-config): restructure package and add configuration files
* Added `LICENSE` and `README.md` files for documentation and licensing.
* Introduced `eslint.config.js` for ESLint configuration.
* Created `src/factory.ts` for defining the base configuration and factory function.
* Updated `package.json` to reflect new structure and added linting scripts.
* Removed obsolete `index.ts` file.
* Adjusted `tsconfig.json` to include only the `src` directory.
