+++ tooling/tsdown-config/eslint.config.js
@@ -1,0 +1,6 @@
+// @ts-check
+import { luxass } from "@luxass/eslint-config";
+
+export default luxass({
+  pnpm: true,
+});
