+++ tooling/tsdown-config/package.json
@@ -1,21 +1,26 @@
 {
   "name": "@ucdjs/tsdown-config",
   "version": "1.0.0",
-  "private": true,
   "type": "module",
+  "private": true,
   "exports": {
-    ".": "./index.ts",
-    "./base": "./base.ts"
+    ".": "./src/factory.ts",
+    "./package.json": "./package.json"
   },
   "files": [
-    "index.ts",
-    "base.ts"
+    "src"
   ],
+  "scripts": {
+    "lint": "eslint .",
+    "typecheck": "tsc --noEmit"
+  },
   "dependencies": {
     "tsdown": "catalog:dev"
   },
   "devDependencies": {
+    "@luxass/eslint-config": "catalog:linting",
     "@ucdjs/tsconfig": "workspace:*",
+    "eslint": "catalog:linting",
     "typescript": "catalog:dev"
   }
 }
